Implemented private route with React Router
