+++ .eslintrc.js
@@ -1,23 +1,23 @@
 module.exports = {
-  env: {
-    browser: true,
-    es2021: true,
-  },
-  extends: ['eslint:recommended', 'plugin:react/recommended', 'plugin:node/recommended'],
-  overrides: [
-    {
-      env: {
-        node: true,
-      },
-      parserOptions: {
-        sourceType: 'script',
-      },
+    env: {
+        browser: true,
+        es2021: true,
     },
-  ],
-  parserOptions: {
-    ecmaVersion: 2021,
-    sourceType: 'module',
-  },
-  plugins: ['react', 'node'],
-  rules: {},
+    extends: ['eslint:recommended', 'plugin:react/recommended', 'plugin:node/recommended'],
+    overrides: [
+        {
+            env: {
+                node: true,
+            },
+            parserOptions: {
+                sourceType: 'script',
+            },
+        },
+    ],
+    parserOptions: {
+        ecmaVersion: 2021,
+        sourceType: 'module',
+    },
+    plugins: ['react', 'node'],
+    rules: {},
 };
