+++ src/pages/ProfilePage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ProfilePage = () => {
+    return (
+        <div>
+            <h2>Protected profile page</h2>
+        </div>
+    );
+};
+
+export default ProfilePage;
